Rename file, expose methods for mocking getEditorNamespace.
